refactored copyright
