+++ README.md
@@ -1,14 +1,5 @@
 # This is my first Material UI React application
 
-## The idea behind the example
-
-<!-- #host-reference -->
-
-This example demonstrates how you can use Material UI with [Create React App](https://github.com/facebookincubator/create-react-app) in [TypeScript](https://github.com/Microsoft/TypeScript).
-It includes `@mui/material` and its peer dependencies, including [Emotion](https://emotion.sh/docs/introduction), the default style engine in Material UI.
-If you prefer, you can [use styled-components instead](https://mui.com/material-ui/integrations/interoperability/#styled-components).
-
-
 ## Why?
 
 I am using this hello world to demonstrate how to use Azure OpenAI to review a pull request.
